smartplan_ai_v3/backend/requirements.txt
# SmartPlan AI v3.0 Backend Dependencies

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Geometry
shapely>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# LLM - MegaLLM (OpenAI-compatible)
openai>=1.0.0

# CAD/DXF
ezdxf>=1.1.0

# Graph Algorithms
networkx>=3.2.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
